♻️ Unify flaw source of truth
